Add files via upload
These are the files for the microservice of camera that is connected to a Raspberry Pi 4. On the Raspberry Pi 4 I have installed the k3s and the Docker. In that docker i have built the "camera_service" microservice which is programmed to take 1 picture using a logitech c505 camera connected to the Raspberry Pi 4 every 10 seconds and send these pictures immediately to another microservice (called the image_database microservice)